Move API response simplification code to a new file
